Added versioning to schemas
The HEPData website uses the older, less strict versions of the schemas from the 0.1.16 tag. We need to allow this whilst keeping the latest version of the validator stricter.
To enable this I've retrieved the old versions and added subdirectories so we can version the schemas.
The validators can now accept a keyword argument 'schema_version' to specify which to use (defaults to v2).
Docs still need to be updated.